Changed the Jenkins specific OutputSink to PlainAnsiOutputSink and removed duplication
